Initial commit of a Vagrant-based VM
Vagrant (and Fabric!) make it a lot easier to work with the uqm
crossbuilder. See the README file for more details. This new
Vagrant-based VM is based on Ubuntu 12.10 and uses Puppet in combination
with a small Fabric script to provision a server.

The advantages of this Vagrant-based VM in a nutshell:
-No manual fiddling with Virtualbox is needed anymore
-Users don't really need to know anything about Linux; Fabric simplifies
things greatly
-Compile UQM using a single command: 'fab build_uqm'
-Integrates nicely with your IDE of choice, as long as your IDE is
capable of running batch files or custom executables
-Modifying the VM is easy and doesn't require a Linux-based setup